fix: clean script

+++ bin/clean.sh
@@ -25,56 +25,80 @@ select opt in "${options[@]}" "ABORT" ; do
 done
 
 echo "Cleaning $(pwd)"
-mv $opt $pwd/trash/client
+if [ -e "$opt" ]; then
+  rm -r $pwd/trash/client/$opt
+  mv $opt $pwd/trash/client
+fi
 
 echo "Going back to:"
 cd -
 
 cd nginx
 echo "Cleaning $(pwd)"
-mv $opt.conf $pwd/trash/nginx
+if [ -e "$opt.conf" ]; then
+  rm -r $pwd/trash/nginx/$opt.conf
+  mv $opt.conf $pwd/trash/nginx
+fi
 
 echo "Going back to:"
 cd -
 
 cd /var/www
 echo "Cleaning $(pwd)"
-mv $opt $pwd/trash/var-www
+if [ -e "$opt" ]; then
+  rm -r $pwd/trash/var-www/$opt
+  mv $opt $pwd/trash/var-www
+fi
 
 echo "Going back to:"
 cd -
 
 cd /etc/nginx/sites-available
 echo "Cleaning $(pwd)"
-mv $opt.conf $pwd/trash/etc-nginx-available
+if [ -e "$opt.conf" ]; then
+  rm -r $pwd/trash/etc-nginx-available/$opt.conf
+  mv $opt.conf $pwd/trash/etc-nginx-available
+fi
 
 echo "Going back to:"
 cd -
 
 cd /etc/nginx/sites-enabled
 echo "Cleaning $(pwd)"
-mv $opt.conf $pwd/trash/etc-nginx-enabled
+if [ -e "$opt.conf" ]; then
+  rm -r $pwd/trash/etc-nginx-enabled/$opt.conf
+  mv $opt.conf $pwd/trash/etc-nginx-enabled
+fi
 
 echo "Going back to:"
 cd -
 
 cd /etc/letsencrypt/archive
 echo "Cleaning $(pwd)"
-mv $opt $pwd/trash/etc-letsencrypt-archive
+if [ -e "$opt" ]; then
+  rm -r $pwd/trash/etc-letsencrypt-archive/$opt
+  mv $opt $pwd/trash/etc-letsencrypt-archive
+fi
 
 echo "Going back to:"
 cd -
 
 cd /etc/letsencrypt/live
 echo "Cleaning $(pwd)"
-mv $opt $pwd/trash/etc-letsencrypt-live
+if [ -e "$opt" ]; then
+  rm -r $pwd/trash/etc-letsencrypt-live/$opt
+  mv $opt $pwd/trash/etc-letsencrypt-live
+fi
 
 echo "Going back to:"
 cd -
 
 cd /etc/letsencrypt/renewal
 echo "Cleaning $(pwd)"
-mv $opt.conf $pwd/trash/etc-letsencrypt-renewal
+if [ -e "$opt.conf" ]; then
+  rm -r $pwd/trash/etc-letsencrypt-renewal/$opt.conf
+  mv $opt.conf $pwd/trash/etc-letsencrypt-renewal
+fi
 
 echo -e '\e[34You might want to stop the pm2 process if it exists\e[39m'
 echo 'pm2 stop $opt'
